requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0
openpyxl>=3.0.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0

